ModelConstructor:  create new file for model lifter
Move `make_*()` from onmt/Models.py.

This file is for models creation, which consults options
and create each Encoder and Decoder accordingly.
